// frontend/src/api/model/githubCom4H1RZooraInternalDomainUserCustomFieldDefinition.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainCustomFieldType } from './githubCom4H1RZooraInternalDomainCustomFieldType';

export interface GithubCom4H1RZooraInternalDomainUserCustomFieldDefinition {
  archived_at?: string;
  created_at?: string;
  description?: string;
  field_type?: GithubCom4H1RZooraInternalDomainCustomFieldType;
  id?: string;
  is_required?: boolean;
  is_unique?: boolean;
  label?: string;
  options?: string[];
  organization_id?: string;
  position?: number;
  updated_at?: string;
  visible_to_user?: boolean;
}